+++ doc/sphinx/changelog.rst
@@ -36,6 +36,18 @@
 :Mach-O:
 
   * Add support for |lief-macho-atom-info| command (``LC_ATOM_INFO``)
+  * Add support for modifying Mach-O commands that embed variable-length data
+    (:issue: `1204`, :issue:`1125`):
+
+    .. code-block:: python
+
+      # Change library name
+      for lib in macho.libraries:
+          lib.name += "/some/path/lib.dylib"
+
+      # Change rpath
+      for rpath in macho.rpaths:
+          rpath.path += '/a/very/long/path/that/needs/expansion'
 
 :ELF: